Fix some bug and add features in main model and main gui

+++ models/main_model.py
@@ -14,7 +14,7 @@ class MainModel:
         self.min_pay = data["minimum payment"]
 
     def update_minimum_pay(self):
-        self.min_pay = int(self.bill * 0.1)
+        self.min_pay = self.bill * 0.1
 
     def purchase(self, amount):
         if amount <= 0:
